--- a/.github/workflows/CI.yml
+++ b/.github/workflows/CI.yml
@@ -13,10 +13,9 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.3'
-          - '1.4'
           - '1.5'
-#          - 'nightly'
+          - '1'
+          - 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest