--- a/Project.toml
+++ b/Project.toml
@@ -2,17 +2,15 @@ name = "AccurateArithmetic"
 uuid = "22286c92-06ac-501d-9306-4abd417d9753"
 license = "MIT"
 repo = "https://github.com/JuliaMath/AccurateArithmetic.jl.git"
-version = "0.3.5"
+version = "0.3.6"
 
 [deps]
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-SIMDPirates = "21efa798-c60a-11e8-04d3-e1a92915a26a"
 VectorizationBase = "3d5dd08c-fd9d-11e8-17fa-ed2836048c2f"
 
 [compat]
-JSON = "0"
-SIMDPirates = "0.2, 0.3, 0.5, 0.6.3, 0.7, 0.8"
-VectorizationBase = "0"
-julia = "1"
+JSON = "0.21"
+VectorizationBase = "0.15"
+julia = "1.5"