--- a/Project.toml
+++ b/Project.toml
@@ -12,5 +12,5 @@ VectorizationBase = "3d5dd08c-fd9d-11e8-17fa-ed2836048c2f"
 
 [compat]
 JSON = "0.21"
-VectorizationBase = "0.15"
+VectorizationBase = "0.19"
 julia = "1.5"