--- a/src/accumulators/compSum.jl
+++ b/src/accumulators/compSum.jl
@@ -8,19 +8,19 @@ end
 
 @inline function add!(acc::CompSumAcc{T, EFT}, x::T) where {T, EFT}
     acc.s, e = EFT(acc.s, x)
-    SIMDops.@fusible acc.e += e
+    @fastmath acc.e += e
 end
 
 @inline function add!(acc::A, x::A) where {A<:CompSumAcc{T, EFT}} where {T, EFT}
     acc.s, e = EFT(acc.s, x.s)
-    SIMDops.@fusible acc.e += x.e + e
+    @fastmath acc.e += x.e + e
 end
 
-@inline function Base.sum(acc::CompSumAcc{T, EFT}) where {T<:Vec, EFT}
+@inline function Base.sum(acc::CompSumAcc{T, EFT}) where {W, T<:Vec{W}, EFT}
     acc_r = compSumAcc(EFT, fptype(T))
     acc_r.e = vsum(acc.e)
-    for xi in acc.s
-        acc_r.s, ei = EFT(acc_r.s, xi.value)
+    for w in 1:W
+        acc_r.s, ei = EFT(acc_r.s, acc.s(w))
         acc_r.e += ei
     end
     acc_r