--- a/src/accumulators/mixedDot.jl
+++ b/src/accumulators/mixedDot.jl
@@ -6,11 +6,11 @@ mixedDotAcc(::Type{Float32})                   = MixedDotAcc(Float64(0))
 mixedDotAcc(::Type{Vec{N, Float32}}) where {N} = MixedDotAcc(vzero(Vec{N, Float64}))
 
 function add!(acc::MixedDotAcc, x, y)
-    SIMDops.@fusible acc.s += SIMDops.f64(x) * SIMDops.f64(y)
+    @fastmath acc.s += f64(x) * f64(y)
 end
 
 function add!(acc::MixedDotAcc{T}, x::MixedDotAcc{T}) where {T}
-    SIMDops.@fusible acc.s += x.s
+    @fastmath acc.s += x.s
 end
 
 Base.sum(acc::MixedDotAcc{T}) where {T<:Vec}  = MixedDotAcc(vsum(acc.s))