--- a/src/accumulators/sum.jl
+++ b/src/accumulators/sum.jl
@@ -5,11 +5,11 @@ end
 sumAcc(T) = SumAcc(vzero(T))
 
 function add!(acc::SumAcc, x)
-    SIMDops.@fusible acc.s += x
+    @fastmath acc.s += x
 end
 
 function add!(acc::SumAcc{T}, x::SumAcc{T}) where {T}
-    SIMDops.@fusible acc.s += x.s
+    @fastmath acc.s += x.s
 end
 
 Base.sum(acc::SumAcc{T}) where {T<:Vec}  = SumAcc(vsum(acc.s))