… src/SIMDops.jl