Fixes #23 

I have to laugh at these `sum` benchmarks (run on my laptop):
```julia
     Testing Running tests...
Test Summary:      | Pass  Total
AccurateArithmetic |  636    636

size 32
  sum_kbn    18.171 ns (0 allocations: 0 bytes)
  sum_oro    24.164 ns (0 allocations: 0 bytes)
  sum_naive  4.052 ns (0 allocations: 0 bytes)
  sum        19.232 ns (0 allocations: 0 bytes)

  dot_oro    26.893 ns (0 allocations: 0 bytes)
  dot_naive  4.716 ns (0 allocations: 0 bytes)
  dot        7.359 ns (0 allocations: 0 bytes)

size 10_000
  sum_kbn    1.243 μs (0 allocations: 0 bytes)
  sum_oro    1.925 μs (0 allocations: 0 bytes)
  sum_naive  480.754 ns (0 allocations: 0 bytes)
  sum        9.890 μs (0 allocations: 0 bytes)

  dot_oro    3.321 μs (0 allocations: 0 bytes)
  dot_naive  962.393 ns (0 allocations: 0 bytes)
  dot        2.198 μs (0 allocations: 0 bytes)

size 1_000_000
  sum_kbn    187.716 μs (0 allocations: 0 bytes)
  sum_oro    246.418 μs (0 allocations: 0 bytes)
  sum_naive  155.506 μs (0 allocations: 0 bytes)
  sum        949.257 μs (0 allocations: 0 bytes)

  dot_oro    572.044 μs (0 allocations: 0 bytes)
  dot_naive  543.026 μs (0 allocations: 0 bytes)
  dot        494.853 μs (0 allocations: 0 bytes)

size 100_000_000
  sum_kbn    46.072 ms (0 allocations: 0 bytes)
  sum_oro    44.406 ms (0 allocations: 0 bytes)
  sum_naive  43.112 ms (0 allocations: 0 bytes)
  sum        107.850 ms (0 allocations: 0 bytes)

  dot_oro    77.514 ms (0 allocations: 0 bytes)
  dot_naive  76.352 ms (0 allocations: 0 bytes)
  dot        70.570 ms (0 allocations: 0 bytes)
     Testing AccurateArithmetic tests passed

julia> versioninfo()
Julia Version 1.7.0-DEV.239
Commit ca075469eb (2021-01-08 18:34 UTC)
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-11.0.0 (ORCJIT, tigerlake)
Environment:
  JULIA_NUM_THREADS = 8
```
Seems that the base pairwise summation algorithm is slower than AccuracteArithmetic's more accurate algorithm, at least for this size * arch combination.