@chriselrod 2021-01-21
Yes, it's ready.