@chriselrod 2021-02-22
There have been a few more VectorizationBase minor version bumps. I'll test and then add them.