@chriselrod 2021-04-10
I just pushed an update for the breaking changes in VectorizationBase 0.19.