@ffevotte 2021-04-10
Thanks! It looks like this also makes the fix in #26 unnecessary.

We should release this new version instead. Are there other changes you'd like to incorporate in v0.3.6, or should I proceed with the release?