@chriselrod 2021-04-10
No, that's all. It'll be nice to finally have a release supporting 1.6.